move SectionItem into svelte module
