+++ client/web-sveltekit/src/lib/search/dynamicFilters/Section.svelte
@@ -1,5 +1,14 @@
+<script lang="ts" context="module">
+    export type SectionItem = Omit<Filter, 'count'> & {
+        count?: Filter['count']
+        selected: boolean
+    }
+</script>
+
 <script lang="ts">
     import { mdiClose } from '@mdi/js'
+
+    import type { Filter } from '@sourcegraph/shared/src/search/stream'
 
     import { page } from '$app/stores'
     import { pluralize } from '$lib/common'
@@ -8,7 +17,6 @@
     import { Badge, Button } from '$lib/wildcard'
 
     import { updateFilterInURL } from './index'
-    import type { SectionItem } from './Section'
 
     export let items: SectionItem[]
     export let title: string
