--- client/web-sveltekit/src/lib/search/dynamicFilters/Section.ts
@@ -1,6 +1,0 @@
-import type { Filter } from '@sourcegraph/shared/src/search/stream'
-
-export type SectionItem = Omit<Filter, 'count'> & {
-    count?: Filter['count']
-    selected: boolean
-}
